windows-mount: moved FsRootNode to own file
